Version 4.0 (#2)
### UIControl
`FloatingRatingView` renamed to `RatingControl`. Also now it is a subclass of `UIControl` instead of `UIView`, because that RatingView is closer to control type than view type.

Now control supports default `UIControl` methods, like `isEnabled`, `addAction`, `@IBAction` outlets and so on. Please use it instead of delegate methods, because it's a more native way.

### Storyboard
You can add storyboard's `@IBAction` outlets in Interface Builder without writing a code for delegate methods. Less code for you.

Also I've added support for `intrinsicContentSize`, so view know about own's desired size (like labels or buttons) and you will need to define less constraints and it will be easier to configure it in Interface Builder.

### Haptic
Support of haptic feedback was added, try that magic on devices with appropriate hardware.

### Compatibility
RatingControl (4.0) is almost compatible with previous version of FloatingRatingView.

1. You need to find and replace import statements because of library renaming:
```swift
import FloatingRatingView // search that
import RatingControl // replace to that
```
2. Minimum required iOS version now is `10.0`.

Because of that 2 breaking changes, version was changed to 4.0.

RatingControl is now a subclass of `UIControl` that's why a lot of old methods was deprecated, you still can use it. XCode will warn you about outdated methods with help messages with fix proposals.

### Other changes
* Support Swift 4.2
* Objective-C support
* Update Example application
* Tidy folder structure